Split up the API and CLI
